chore: Add experimental custom server middleware
