+++ packages/nuxt/src/module.ts
@@ -1,7 +1,21 @@
 import { addComponent, addImports, addPlugin, addRouteMiddleware, addServerHandler, addTypeTemplate, createResolver, defineNuxtModule, updateRuntimeConfig } from '@nuxt/kit'
 import type { IsomorphicClerkOptions } from 'vue-clerk'
 
-export type ModuleOptions = Omit<IsomorphicClerkOptions, 'routerPush' | 'routerReplace'>
+export type ModuleOptions = Omit<IsomorphicClerkOptions, 'routerPush' | 'routerReplace'> & {
+  /**
+   * Disable the server middleware. This requires you to handle the middleware yourself.
+   *
+   * @example
+   * ```typescript
+   * import { clerkMiddleware } from 'vue-clerk/server'
+   *
+   * export default clerkMiddleware((event) => {
+   *   console.log('auth', event.context.auth)
+   * })
+   * ```
+   */
+  __experimental_disableServerMiddleware?: boolean
+}
 
 export default defineNuxtModule<ModuleOptions>({
   meta: {
@@ -45,10 +59,13 @@ export default defineNuxtModule<ModuleOptions>({
 
     addPlugin(resolver.resolve('./runtime/plugins/clerk'))
 
-    addServerHandler({
-      middleware: true,
-      handler: resolver.resolve('./runtime/server/middleware'),
-    })
+    if (!options.__experimental_disableServerMiddleware) {
+      addServerHandler({
+        middleware: true,
+        handler: resolver.resolve('./runtime/server/middleware'),
+      })
+    }
+
     addServerHandler({
       route: '/api/_clerk/current-user',
       handler: resolver.resolve('./runtime/server/api/current-user.get'),
