+++ packages/nuxt/src/runtime/server/index.ts
@@ -1,3 +1,4 @@
 export * from '@clerk/backend'
 export { clerkClient } from './clerkClient'
 export { getAuth } from './clerkClient'
+export { clerkMiddleware } from './clerkMiddleware'
